Change layout of rust contract
