+++ templates/rust/contract/src/main.rs
@@ -1,73 +1,32 @@
+//! Generated by capsule
+//!
+//! `main.rs` is used to define rust lang items and modules.
+//! See `entry.rs` for the `main` function. 
+//! See `error.rs` for the `Error` type.
+
 #![no_std]
 #![no_main]
 #![feature(lang_items)]
 #![feature(alloc_error_handler)]
 #![feature(panic_info_message)]
 
-// Import from `core` instead of from `std` since we are in no-std mode
-use core::result::Result;
+// define modules
+mod entry;
+mod error;
 
-// Import heap related library from `alloc`
-// https://doc.rust-lang.org/alloc/index.html
-use alloc::{vec, vec::Vec};
-
-// Import CKB syscalls and structures
-// https://nervosnetwork.github.io/ckb-std/riscv64imac-unknown-none-elf/doc/ckb_std/index.html
 use ckb_std::{
-    entry,
     default_alloc,
-    debug,
-    high_level::{load_script, load_tx_hash},
-    error::SysError,
-    ckb_types::{bytes::Bytes, prelude::*},
 };
 
-entry!(entry);
+ckb_std::entry!(program_entry);
 default_alloc!();
 
-/// Program entry
-fn entry() -> i8 {
+/// program entry
+fn program_entry() -> i8 {
     // Call main function and return error code
-    match main() {
+    match entry::main() {
         Ok(_) => 0,
         Err(err) => err as i8,
     }
 }
 
-/// Error
-#[repr(i8)]
-enum Error {
-    IndexOutOfBound = 1,
-    ItemMissing,
-    LengthNotEnough,
-    Encoding,
-    // Add customized errors here...
-}
-
-impl From<SysError> for Error {
-    fn from(err: SysError) -> Self {
-        use SysError::*;
-        match err {
-            IndexOutOfBound => Self::IndexOutOfBound,
-            ItemMissing => Self::ItemMissing,
-            LengthNotEnough(_) => Self::LengthNotEnough,
-            Encoding => Self::Encoding,
-            Unknown(err_code) => panic!("unexpected sys error {}", err_code),
-        }
-    }
-}
-
-fn main() -> Result<(), Error> {
-    // remove below examples and write your code here
-
-    let script = load_script()?;
-    let args: Bytes = script.args().unpack();
-    debug!("script args is {:?}", args);
-
-    let tx_hash = load_tx_hash()?;
-    debug!("tx hash is {:?}", tx_hash);
-
-    let _buf: Vec<_> = vec![0u8; 32];
-
-    Ok(())
-}
